requirements.txt
jupyterlab>=3.5.0
matplotlib>=3.6.2
numpy>=1.22.3
pandas>=1.5.1
scikit-learn>=1.2.0

autograder-py>=0.6.17
